backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
boto3==1.18.44
python-dotenv==0.19.0
openai==0.27.0

